Make wesocket connection handling more robust.
This commit means that the last request for the WebSocket connection
will now get it and will notify an existing connection that it is
being disconnected. This should stop the occasional error where the
WebSocket pool filled up and wouldn't work. It also handles multiple
parties trying to connect more explicitly rather than just failing.